+++ src/editor/console/__tests__/console-pane.test.js
@@ -1,10 +1,10 @@
 import { shallow } from "enzyme";
 import React from "react";
 
-import HistoryItem from "../history-item";
+import HistoryItem from "../history/components/history-item";
 
 import { ConsolePaneUnconnected, mapStateToProps } from "../console-pane";
-import OnboardingContent from "../../../../shared/components/onboarding-content";
+import OnboardingContent from "../../../shared/components/onboarding-content";
 
 describe("ConsolePaneUnconnected React component", () => {
   let props;
